% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appraisalsummary.R
\name{appraisalsummary}
\alias{appraisalsummary}
\title{appraisalsummary}
\usage{
appraisalsummary(myin, simu = FALSE)
}
\arguments{
\item{myin}{A list with the result generated by one of the appraisal functions}

\item{simu}{A parameter that is FALSE (the appraisal is derived from the distribution) or TRUE (the appraisal is derived by simulation)}
}
\value{
nothing -- the function prints
}
\description{
Outputs the results of the appraisal generated by the appraisal functions
}
\examples{
appraisalsummary(myin)
}
