% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_spatial.R
\name{e_spatial}
\alias{e_spatial}
\title{e_spatial}
\usage{
e_spatial(w, lambda, u, tst = 10^(-7), nmax = 50)
}
\arguments{
\item{w}{A spatial weight matrix (a sparce matrix)}

\item{lambda}{The spatial weight parameter}

\item{u}{A vector of iid normal random terms}

\item{tst}{Criterion for stopping the iterations (default: 1E-7)}

\item{nmax}{maximum number of iterations (default: 50)}
}
\value{
A vector of spatially structured random variables
}
\description{
Generates spatially structured error terms for appraisal simulation
}
\details{
This function generates spatially structured error terms for appraisal simulation.
The spatial structure is provided by the spatial weight matrix parameter w, the
strength of the spatial influence by the parameter lambda. The function generates
ONE vector of spatially structured error terms from ONE vector of iid normal
inputs (parameter u).

Parameters tst and nmax can be set to guide the number of iterations in the function.
Smaller values of tst lead to more precise results, but also longer computation.
}
\examples{
w <- listw2sparse(nm_listw)
eps <- rnorm(1953,0,1)
tst <- e_spatial(w, 0.5, eps, tst = 10^(-10))
}
