% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listw2sparce.R
\name{listw2sparse}
\alias{listw2sparse}
\title{listw2sparce}
\usage{
listw2sparse(nm_listw, force = FALSE)
}
\arguments{
\item{nm_listw}{A listw object}

\item{force}{A logical variable that forces computation (default: FALSE)}
}
\value{
A sparce matrix
}
\description{
Converts a listw object (spatial weight matrix) to a sparce matrix.
}
\details{
The function checks the size of the input listw object. If it contains
more than 50,000 elements, the function aborts and issues a notification.
This is typically the case with distance band based spatial weight matrices.
To force the function to run anyways, set the parameter force to TRUE.
}
\examples{
w <- listw2sparse(nm_listw)
}
