% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_quality.R
\name{model_quality}
\alias{model_quality}
\title{model_quality}
\usage{
model_quality(reg)
}
\arguments{
\item{reg}{A regression object}
}
\value{
Printed
}
\description{
A function to summarize indicators of model quality of a regression model
}
\examples{
model_quality(reg)
}
