% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{plot_residuals}
\usage{
plot_residuals(reg)
}
\arguments{
\item{reg}{}
}
\value{
plot A reg object generated by lm()
}
\description{
Plots the histogram of the observed residuals and the corresponding normal density
}
\details{
This function gets the residuals from the reg object and extracts mean and
standard deviation therefrom. It then plots a hastogram of the observed
residuals of the regression. It overlays the histogram with the curve of
a normal density with the mean and standard deviation of the residuals.
}
\examples{
plot_residuals(reg)
}
