% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_inv.R
\name{w_inv}
\alias{w_inv}
\title{w_inv}
\usage{
w_inv(m, lambda)
}
\arguments{
\item{m}{The matrix with the last row of the powers of the spatial weight matrix}

\item{lambda}{A value for the spatial weight parameter}
}
\value{
The approximate inverse of the matrix (I-lambda*w).
}
\description{
Calculates the approximate inverse of the matrix (I-lambda*w).
}
\details{
The function uses the output of the function w_powers and a simulated value for
lambda to calculate the approximate inverse of the matrix (I-lambda*w).
}
\examples{
tst <- w_powers(w, max=5)
inv <- w_inv(tst, 0.5)
}
