% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_powers.R
\name{w_powers}
\alias{w_powers}
\title{w_powers}
\usage{
w_powers(w, max = 5)
}
\arguments{
\item{w}{A spatial weight matrix (A sparce matrix)}

\item{max}{The number of powers to calculate}
}
\value{
A matrix with max+1 rows and the same number of columns as w
}
\description{
Generates powers of the spatial weight matrix (for approximate matrix inverse)
}
\details{
This function generates the powers of the spatial weight matrix (parameter w)
as input for the function w_inv that calculates the approximate inverse. The function
returns the values for the last row (the property to appraise) of the respective
matrices.
}
\examples{
m <- w_powers(w)
}
